Update LinkedInt.py
Fixed a small bug when location data wasn't populated the script would error out. Converted to try/except like some of the other data variables. Added in a line to silence the URL warnings and also cleaned up some code that was printing out the session data so it's cleaner for screenshots.